[fix] (privilege) fix bugs of check multi roles privilege ## Case
when wanted privs is in different roles, single role is not satisfied,but all roles may satisfy.

example: the role1 has select_priv, the role2 has grant_priv, the wanted privs are select_priv && grant_priv,neither role1 nor role2 is satisfied, but (role1 && role2) is satisfied.

 so can not check by for-loop simply

## Solution

When the required permissions are AND, split the permissions and perform the checks. After all role checks are completed, determine whether the user has all the required permissions.